pyproject.toml
[tool.poetry]
name = "BotMyDesk"
version = "v1"
description = "Slackbot for BookMyDesk"
license = "MIT License"
readme = 'README.md'
repository = "https://github.com/dennissiemensma/slack-botmydesk"
homepage = "https://github.com/dennissiemensma/slack-botmydesk"
authors = ["Dennis Siemensma"]


[tool.poetry.dependencies]
python = "^3.10"

coreapi = "^2.3"
django = "^4.0"
django-debug-toolbar = "^3.2"
gunicorn = "^20.1"
platformdirs = "^2.0"
python-decouple = "^3.5"
pytz = "^2022.1"
requests = "^2.26"
slack_sdk = "^3.18"
urllib3 = "^1.26"

[tool.poetry.dev-dependencies]
flake8 = "^4.0"
flake8-bandit = "^3.0"
flake8-black = "^0.3"
flake8-bugbear = "^22.4"
polib = "^1.1"
pytest-cov = "^3.0"
pytest-django = "^4.4"
pytest-xdist = "^2.4"
safety = "^1.10.3"
mypy = "^0.971"
types-pytz = "^2021.0"
types-requests = "^2.0"
types-polib = "^1.0"
types-six = "^1.0"
